+++ cmd/skaffold/app/cmd/flags.go
@@ -24,7 +24,12 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
+	"github.com/GoogleContainerTools/skaffold/cmd/skaffold/app/flags"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/constants"
+)
+
+var (
+	fromBuildOutputFile flags.BuildOutputFileFlag
 )
 
 // Flag defines a Skaffold CLI flag which contains a list of
@@ -422,6 +427,15 @@ var flagRegistry = []Flag{
 		DefValue:      false,
 		FlagAddMethod: "BoolVar",
 		DefinedOn:     []string{"build", "debug", "delete", "deploy", "dev", "run"},
+	},
+	{
+		Name:          "build-artifacts",
+		Shorthand:     "a",
+		Usage:         "File containing build result from a previous 'skaffold build --file-output'",
+		Value:         &fromBuildOutputFile,
+		DefValue:      "",
+		FlagAddMethod: "Var",
+		DefinedOn:     []string{"deploy"},
 	},
 }