+++ pkg/skaffold/docker/client.go
@@ -21,7 +21,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"os/exec"
 	"path/filepath"
 	"sort"
 	"strings"
@@ -204,17 +203,6 @@ func getMinikubeDockerEnv(minikubeProfile string) (map[string]string, error) {
 			return nil, fmt.Errorf("unable to parse minikube docker-env keyvalue: %s, line: %s, output: %s", kv, line, string(out))
 		}
 		env[kv[0]] = kv[1]
-	}
-
-	if found, _ := util.DetectWSL(); found {
-		// rewrite Unix path to Windows
-		cmd := exec.Command("wslpath", env["DOCKER_CERT_PATH"])
-		out, err := util.RunCmdOut(cmd)
-		if err == nil {
-			env["DOCKER_CERT_PATH"] = strings.TrimRight(string(out), "\n")
-		} else {
-			return nil, fmt.Errorf("can't run wslpath: %s", err)
-		}
 	}
 
 	return env, nil